git: Improve exception discipline and error messages.
We use a mix of error codes and raw exceptions in the git updater.

Fix this to use the exception objects used elsewhere (and which give
better error messages).

This also simplifies the code by removing error code returns from some
internal functions whose return value is then just converted into an
exception anyways. Easier just to throw the exception directly and
simplify the higher-level functions.

The one nuance here is that installing a git snapshot can fail yet that
should not cause an exception, we'd just fallback to doing a normal
git-clone.