build-systems: Mask default options for non-default build systems.
This is a first-step hack to have non-default build systems ignore
global options which are related to the build system in use (e.g.
compile or make flags, install/uninstall options, etc.).

This should fix bug 331654 (now that I understand it better) and improve
the situation for bug 332789 (which would be the long-term fix when
implemented).

BUG:331654
CCBUG:332789
FIXED-IN:1.16