Make make-install-prefix a common option again.
I had made some build-system related options not automatically apply to
modules with customized build systems (e.g. Qt, CMake, "custom"). But
this seems like the kind of option we'd want to have in common, as noted
in a bug report to me by Andre Wöbbeking, so make make-install-prefix
apply to all build systems by default if it's a global option.

This can, of course, still be overridden for individual modules or
module-sets.

Original commit: 4a949962
https://invent.kde.org/sdk/kdesrc-build/-/commit/4a949962c6bf45876769e7a4f1048ebe95d1b756