Improve first-run behavior with build metadata.
This commit makes the kde-build-metadata module a requirement instead of
an option, since the vast majority of runs will require this module now.

Additionally the --metadata-only option is added and documented to allow
for downloading the kde-build-metadata module alone (and make the
--pretend option work afterwards), and kdesrc-build recommends using it
if you run with --pretend and without metadata.

This should hopefully make the first-run use case easier for users.

BUG:337446
FIXED-IN:1.16

Original commit: 6911da5c
https://invent.kde.org/sdk/kdesrc-build/-/commit/6911da5c8713c3e33e8e43c6707a2c871bf79375