Add new option 'include-dependencies'.
Currently it (probably) doesn't support pulling in options for these
modules from their host module-sets (although using 'options' sections
might actually work). I think it ignores the 'ignore-modules' option as
well.

Other than that, it should do pretty much what it says on the tin,
although I had to implement a few other bug fixes as well to get it to
that point.

I want to say that there was an open bug for this but I can't seem to
find it now.

Original commit: b452401f
https://invent.kde.org/sdk/kdesrc-build/-/commit/b452401f50d3c740e8529986bf19e074920be0c6