git: Check for in-progress merges or rebases before updating existing repo.
There seems to be nothing better than checking for the existence of
files or directories under .git at this point (though it's possible to
decode git-status --porcelain output without too much trouble in the
case of in-progress merges).

This should hopefully help avoid kdesrc-build trying to update a module
that is being worked on.

BUG:334324
FIXED-IN:1.16