dep-resolv: --include-dependencies; cmdline selections override rcfile.
Add and document the --include-dependencies option. Useful with
--print-modules as a quick way to determine what dependencies a given
module on the command line has, especially if you do a quick
"kdesrc-build --metadata-only" first.

Also, make selecting modules on the command line override
"include-dependencies" as an rc file option, since otherwise it's
impossible to ask just to build a single module if that module is in a
module set with included dependencies.

Use the newly-added --include-dependencies option to force dependency
resolution to happen even if you specify modules to build on the command
line.