kdepim-build-include: Simplify decls.
It is not necessary to name module-sets, that's just a convenience, so
remove names from single-module module-sets.

Also while kdesrc-build won't complain if the set's name matches a
module name (or path), that's not really a good idea, so try to make
remaining module-set names distinct.

Note this is a possible behavior change, if you got used to
"kdesrc-build kdepim" building a bunch of other things in addition to
kde/kdepim. My apologies but many users just want to build kdepim the
single repo, you can always copy the old file into a customized
kdesrc-buildrc.