Add (and document) a --query flag, to print information about modules.
This still takes too long for most shell shortcuts, but will make it
possible to do things like setup shell functions to cd right to a build
dir based only on a module name, check for which git branch a module is
resolved to, or see what config file options a module ended up falling
under.

As part of the documentation changes, I also ported the docs to KF5.