kde-projects: Fix a bug where some cmdline options were ignored.
Though I admit that I don't yet fully understand why the bug was
present, this seems to fix the use of options like --src-only. There was
an actual bug here: command line options are supposed to override
rc-file options by default, so the call to update the global context
object's options had to happen after the rc-file options were read.

Whatever *this* the problem was, it appears to have been related to
either the generation of the list of potential modules, or the filtering
of that list aftewards.

If other problems are encountered please reopen.

BUG:370534
FIXED-IN:16.11