purpose and kirigami are not frameworks -> revert.
purpose is in playground/libs. Apparently it's needed by kamoso, so it should be moved to extragear/libs and added to kf5-extragear-build-include. Follow the review process for this to happen.

kirigami is in extragear/libs. If it's required by discover, it should be moved to kde/workspace, and then it can be added in kf5-workspace-build-include.
If it's also needed by applications, then it needs to be turned into a proper framework....

REVIEW: 129117
CCMAIL: yurchor@ukr.net, aleixpol@kde.org, notmart@gmail.com