ipc: Show status of unbuilt-but-updated modules in async.
Basically, modules which are updated by default but not actually built
(e.g. due to using the manual-build option) are never mentioned in the
kdesrc-build output in the default async mode of operation.

This required a bit of refactoring of the IPC code to fix, since the
underlying cause is that the output during async mode is driven by the
needs of the build thread, which would request updates from the monitor
thread on a per-module basis.  But the build thread never inquired about
modules only being updated, so any such modules were never mentioned on
the command line.