Reduce importance of qtdir option, default to ''.
The qtdir setting took on much less importance with Qt 5.  We've never
officially supported building Qt 5 with kdesrc-build, and qmake and
cmake both use qmake or similar Qt-provided config information to find
Qt, rather than using the unsupported QTDIR environment variable.

As a result, default the value to an empty string (interpreted as use of
system Qt) and stop adding it automatically to the environment variables
defined during the build.  If set, it will continue to be used for now.

This also helps the environment setup driver detect when it should look
for the system Qt5 qmake.

Original commit: 02c616cb
https://invent.kde.org/sdk/kdesrc-build/-/commit/02c616cbb08915fe2804c75259ddf449ab8c1792