Ensure all potential logdirs setup prior to async execution.
The log directories are stored in a central directory in the
BuildContext so that we know which ones we're supposed to reuse.  But
this only happens on first use.  If this first use happens during an
async build it won't be properly persisted, so each succeeding build
will think it's the first one to need a log directory.

This can happen for modules that use a log directory different from the
global directory.