kde-projects: Track how a module became a selection candidate.
This is needed to try to perform filtering of the module list based on
which git branches are available, which we could do easily when using
kde_projects.xml.

Without that data, we effectively have to wait until the module list is
already made and about to git update a module before we can tell if the
branch will be there.  By that time it's too late to have filtered the
module list, so we have to determine whether this should be a hard error
or not later in execution.

To make that choice, we essentially need to know whether the module is
being built because it was specifically asked for or whether it was
brought in as part of a larger module set (and therefore can be ignored
if the git branch is missing).

This adds tracing for the config file to supplement existing tracing of
cmdline selectors.