kde-projects: Fix runtime check for YAML modules.
My attempt to make the KDEProjectReader's check for YAML be a runtime check was
thwarted because I had separately used a "use KDEProjectReader" elsewhere,
which Perl ensures happens at compile time.

Instead make the check part of the constructor so that the package can at least
compile without throwing an error.