persistence: Fix regression in saving options.
The move to JSON storage of persistent opts break first-time users,
since the refactored code never actually gets around to setting a
persistent_options key in the build context object.

This seemed to work previously with Data::Dumper, but JSON::PP is not
happy with being asked to serialize an undef.  On the other hand I had
always thought that sub setPersistentOption would have 'autovivified'
the needed hash table, and the reporter's build process should have
resulted in many calls to setPersistentOption over the build.

BUG:388180
FIXED-IN:18.01