Don't shortcut build output for non-CMake build systems.
There is code to avoid using the TTY code that updates the status of the
build in the current line.  This code prevents using the fancy TTY stuff
if progress output is not supported, but doesn't actually ever output
the end result of the build (success or failure), which wasn't the
intent.

In the other situations we don't want to clutter the line output with a
redundant status.  Instead of duplicating the "progress output
supported" check later to output the status (or not), I just took out
the check for whether progress output is supported entirely, since the
progress callback can work even with build systems w/out support.

BUG:391396
FIXED-IN:18.04