mojo: Consistently use promises/continuation-passing.
The current attempt is a mix of imperative code (that tries to manually
wait for promises to resolve) and promises.  Mojolicious (and JS
promises and/or continuation-passing style codes in general) really
don't work as well in this mode, so bite the bullet and try to
consistently use promises and promise-style code in this section.

In other words, start off with all the imperative setup, switch over to
and consistently use promises and closure-based callbacks, and only once
that is /all/ done do we switch back to an imperative mode.

I may still be missing some things but that will have to be the approach
going forward.