mojo: Handle modules whose updates were skipped.
The old IPC stuff had code to specially handle when modules were not
updated (for whatever reason), and this was commingled with the build
code.

This is fixed by using shell Promises for modules in the build list but
not the update list, that determine whether to proceed with the build
for that module based on the criteria we used earlier (essentially,
build if it failed before, or if we're in --no-src).

I feel like there's still a better way to handle this though.