mojo: Re-add the status server.
This resurrects the basic code I had in the old _handle_monitoring.

This still needs some polishing but seems to work! :)

But next I need to figure out what kind of information needs to be
available in the status server and how best to track it, but at least
the basic infrastructure is there now (that is, if you have Mojolicious
installed for Perl...).