mojo: Complete promise-ifying Module.pm's build subroutine.
One somewhat annoying part about this is that the Mojo subprocess stuff
has no event loop running at all in portion of code that actually runs
in the forked child.  So if you want to use promise-based code within a
subprocess block (to allow for doing build/test/install in a single
subprocess), you need to break up each chunk into its own subprocess.

There is probably a way to do this (Mojo::Reactor->next_tick?) but it's
hard to tell what's a documented 'feature' in the Mojo API or whether
I'd be using an undocumented behavior that can break in the future.