git: Ensure we skip modules that we say we must skip.
The git code recognizes when a module that is selected for update/build
implicitly (e.g. because it is included as a subset of a module grouping
that *was* asked for) does not actually have a git branch.  This often
happens for newly-introduced modules (which obviously won't already have
historical "stable" branches).

Rather than forcing users to know that a module is new and manually
ignore it we announce we will skip the build for these modules.

However I confused the return code in this function; no errors are
returned, only exceptions are used for errors, which meant kdesrc-build
would try to merrily build the non-existant module anyways.

Fixed by making this an error.  I looked at making it possible to
completely elide the module from the update/build list entirely but that
requires a network call (we can't rely on cached source code since
we can't download non-existent sources), and I don't think that's
a good idea.  The error message can probably be worded better though.

BUG:392654
FIXED-IN:18.04