mojo: Use daemon server instead of rolling our own.
This is both better-supported by Mojolicious documentation and reduces
code complexity.

The only reason I was trying to manually handle requests before is that
I took the Mojolicious documentation to mean that Mojo::Server::Daemon
automatically would fork a new server process, which would not play well
with the existing multi-process code base.

But it turns out it doesn't do this, the "daemon" seems to merely mean
that the code is ready to run as a separate daemon but it does not do
this by itself and there's no requirement to do so.