mojo: Apply module option changes that happen in a subprocess.
This requires use to record option changes and return them back to the
parent process so that they can be replayed.

Doing this allows for things like setting the module option to save the
location of the error logfile to work.  Currently there is a built-in
fallback to just assume that there's a relevant error.log, which means
most of the time no one had reason to suspect that Module::setOption
doesn't work right during subprocesses.

This isn't specific to the Mojolicious-based changes (log_command has
been broken in this respect for a long time...) but fixing this seemed
necessary for some other status viewer improvements.