mojo: Try to hyperlink error log files if errors happen.
My initial attempt was just to give the file:// URL but browser security
restrictions seem to make this untenable since the main status page is
provided over http://.

The workaround I implemented is to have kdesrc-build implement providing
the file to the browser, but that only works for as long as the build is
happening.

I have some modules for testing which fail without causing a log file as
well, so I still need to come up with something better for that use
case.