Don't double-set CMAKE_PREFIX_PATH if set by user.
This fixes bug 395627, where the error is actually that we set
CMAKE_PREFIX_PATH twice if qtdir is set to a non-system path and the
user is also setting CMAKE_PREFIX_PATH.

Unfortunately the second value overrides the first (the one the user
set).

Also added a test for this, which fails before the fix and passes
afterwards. The full test suite (all 5...) pass.

BUG:395627
FIXED-IN:18.08