mojo: Fix position of module names in progress.
This keeps the position of the module name being built from moving around
rapidly when the module being updated changes.

This also has the effect of actually outputting to the maximum possible width,
which showed that I had to account for more of the extra space in the output of
the overall line to fit everything in.