mojo: Move phase-handling code to a more central spot.
Right now the dependencies between the module build phases, and the
actions to take in each phase, are all commingled together.

I want to break those up so that *how* a phase is conducted is
independently of *whether* (and in what order) a phase is run. I've
wanted this since ksb::PhaseList was defined in fact, just had no way to
easily express that.

But with Mojolicious I can make the phase handlers mere promises, and
then separately link the handlers together, or not, as appropriate for
each build run.