mojo: Add helper class to create "promise chains".
The promise chain concept I have would allow for promises (in the JavaScript
and Mojolicious sense, think async/await) to be generated once at the script
start, and then control flow of execution in the proper order, safely, with
maximum concurrency feasible, and all for nearly "free".

It took me longer than I want to admit to get something working but I have test
code using the concept that works now, and next it's a matter of integrating
the existing promise-using code into this.