mojo: Add a backend-support class to allow headless ops.
This adds a backend support class (BackendServer, no ksb:: prefix for
now since Mojolicious uses the basedir for this module to look for
supporting files, like the HTML/embedded Perl templates).

To use the backend, for now run kdesrc-build with only the option
--backend, which will start the server and output the URL to stdout.

This introduces another class as well (ksb::UserInterface::TTY) to serve
as the default user interface, and makes kdesrc-build the driver for
this interface (instead of the more generic ksb::Application). This uses
the backend interface directly as well, enforcing its normal use.

There's a lot more to do to refactor a clean separation between
model/controller and the user interface / "view" here, especially at the
TTY, but this code does at least work and I think it's a clean enough
foundation that all the heavy lifting should be done.